project page component:
- Created project page component
- updated homepage
